Documentation and tests for TrimmedLineDiff
Refactoring to avoid changing tokenize API, adding documentation to
README.md, and adding 4 tests for TrimmedLineDiff to diffTest.js.

I also found a bug in how TrimmedLineDiff handled windows new lines
while adding the test, as well as changed TrimmedLineDiff so it doesn't
add a newline character ('\n') if it is the last line.