Tweak eslint rules (#594)
* Remove unneeded disabling of no-inner-declarations

No reason to have this turned off; we don't violate the rule anywhere and I don't think I've ever worked on a codebase that did violate it tbh

* Tweak comment to explain why no-unused-expressions needs disabling in tests

* Remove lines setting rules to 0 - they don't do anything

Maybe in an old version of eslint, these disabled on-by-default rules, but there ARE no on-by-default rules in the latest version of eslint, so these lines have no effect. Seems like deleting them all simplifies the config.

* Remove more unnecessary setting of rules to 0, and a couple of stray console.log lines