Remove no-longer-needed npm-debug.log entry from .gitignore (#597)
Reasons this is no longer needed:
- all the instructions for contributing this repo say to use yarn, which has never generated npm-debug.log
- since npm 4.2.0, npm *also* doesn't generate npm-debug.log

I noticed this while cross-checking .gitignore against .npmignore to make sure there weren't any weird generated files, not meant to be published, that we were Git-ignoring but forgetting to put in .npmignore. This was the only such inconsistency. We could resolve it by adding the same line to .npmignore, but since npm-debug.log will probably never get generated on anyone's machine ever again while they're working on jsdiff, it's probably better to just remove all reference to it.